fuse: support idmapped mounts
For now, only works when posix-acl/access-control xlator is disabled for the volume.

Reason for that is that this xlator is a server-side and performs all the permission
checks on the server side, which is very-very problematic from the idmapped mounts perspecive.
It's important to notice, that while disabling "features.acl" is not recommended and considered
as a debugging option only.
At the same time, we don't disable permission checks completely by doing that, because
when fuse client is used permission checks are done on the kernel side ("default_permissions" fuse mount option).

But anyways, this is not a final and ready-to-go support for idmapped mounts anyways. This is
a PoC example and good point to start discussion on the problems and approaches.

The testing procedure:

1. Modify setup.sh like this to disable features.acl (this disables posix-acl/access-control xlator):
--- a/tests/bugs/fuse/setup.sh
+++ b/tests/bugs/fuse/setup.sh
@@ -12,9 +12,10 @@ TEST $CLI volume info

 #Create a distributed volume
 TEST $CLI volume create $V0 $H0:$B0/${V0}{1..2};
+TEST $CLI volume set $V0 features.acl false
 TEST $CLI volume start $V0

2. make && make install
3. ./tests/bugs/fuse/setup.sh

4. ./mount-idmapped --map-mount b:1000:0:2 /mnt/glusterfs/0 /idmapped
[ taken from https://raw.githubusercontent.com/brauner/mount-idmapped/master/mount-idmapped.c ]

5. *play with /idmapped mount*

6. ./tests/bugs/fuse/teardown.sh

Some extra links and examples:

- kernel support
https://github.com/mihalicyn/linux/commits/fuse_idmapped_mounts

- libfuse support
https://github.com/mihalicyn/libfuse/commits/idmap_support

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>